sprin-5100 adding service to update and create lucy operations for all remaining entities. Also did refactoring
