+++ daaily/lucy/resources/group/resource.py
@@ -1,21 +1,21 @@
 from typing import Any, Dict, Generator
 
-from daaily.lucy.enums import EntityType
+from daaily.lucy.enums import EntityType, Service
 from daaily.lucy.models import Filter
 
-from . import BaseResource
+from .. import BaseResource
 
 
-class SpacesResource(BaseResource):
+class GroupsResource(BaseResource):
     def get(
         self, filters: list[Filter] | None = None
     ) -> Generator[Dict[str, Any], None, None]:
         """
-        Retrieves spaces with optional filtering, returning them as a generator
-        that yields each space one at a time.
+        Retrieves groups with optional filtering, returning them as a generator
+        that yields each group one at a time.
 
         Available filters:
-            - space_ids (str): Filter by comma separated space IDs.
+            - group_ids (str): Filter by comma separated group IDs.
 
         Note that the following filters are automatically added to the query:
             - skip (int): Number of records to skip.
@@ -30,14 +30,14 @@ class SpacesResource(BaseResource):
         Example:
             ```python
             # Define filters
-            filters = [Filter("space_ids", "12345, 78910")]
+            filters = [Filter("group_ids", "12345, 78910")]
 
-            # Get spaces (pagination handled internally)
-            spaces = client.spaces.get(filters=filters)
+            # Get groups (pagination handled internally)
+            groups = client.groups.get(filters=filters)
 
             # Iterate over the results without worrying about pagination
-            for s in spaces:
-                print(f"ID: {s['space_id']}, Space Type: {s['space_type']}")
+            for g in groups:
+                print(f"ID: {g['group_id']}, Name: {g['name_en']}")
             ```
         """
         if filters is None:
@@ -57,7 +57,7 @@ class SpacesResource(BaseResource):
         new_filters.append(limit_filter)
         new_filters.append(skip_filter)
         while True:
-            response = self._client.get_entities(EntityType.SPACE, new_filters)
+            response = self._client.get_entities(EntityType.GROUP, new_filters)
             if response.status != 200:
                 break
             for item in response.json():  # type: ignore
@@ -67,11 +67,25 @@ class SpacesResource(BaseResource):
             new_filters = [f for f in new_filters if f.name != "skip"]
             new_filters.append(skip_filter)
 
-    def get_by_id(self, space_id: int):
-        return self._client.get_entity(EntityType.SPACE, space_id)
+    def get_by_id(self, group_id: int):
+        return self._client.get_entity(EntityType.GROUP, group_id)
 
-    def update(self, spaces: list[dict], filters: list[Filter] | None = None):
-        return self._client.update_entities(EntityType.SPACE, spaces, filters)
+    def update(
+        self,
+        groups: list[dict],
+        filters: list[Filter] | None = None,
+        service: Service = Service.SPARKY,
+    ):
+        return self._client.update_entities(
+            EntityType.GROUP, groups, filters, service=service
+        )
 
-    def create(self, spaces: list[dict], filters: list[Filter] | None = None):
-        return self._client.create_entities(EntityType.SPACE, spaces, filters)
+    def create(
+        self,
+        groups: list[dict],
+        filters: list[Filter] | None = None,
+        service: Service = Service.SPARKY,
+    ):
+        return self._client.create_entities(
+            EntityType.GROUP, groups, filters, service=service
+        )
