+++ daaily/lucy/resources/journalist/resource.py
@@ -1,21 +1,21 @@
 from typing import Any, Dict, Generator
 
-from daaily.lucy.enums import EntityType
+from daaily.lucy.enums import EntityType, Service
 from daaily.lucy.models import Filter
 
-from . import BaseResource
+from .. import BaseResource
 
 
-class GroupsResource(BaseResource):
+class JournalistsResource(BaseResource):
     def get(
         self, filters: list[Filter] | None = None
     ) -> Generator[Dict[str, Any], None, None]:
         """
-        Retrieves groups with optional filtering, returning them as a generator
-        that yields each group one at a time.
+        Retrieves journalists with optional filtering, returning them as a generator
+        that yields each journalist one at a time.
 
         Available filters:
-            - group_ids (str): Filter by comma separated group IDs.
+            - journalist_ids (str): Filter by comma separated journalist IDs.
 
         Note that the following filters are automatically added to the query:
             - skip (int): Number of records to skip.
@@ -25,19 +25,19 @@ class GroupsResource(BaseResource):
             filters (list[Filter] | None): A list of filters to apply to the query.
 
         Yields:
-            dict: A dictionary representing a single space.
+            dict: A dictionary representing a single journalist.
 
         Example:
             ```python
             # Define filters
-            filters = [Filter("group_ids", "12345, 78910")]
+            filters = [Filter("journalist_ids", "12345, 12322")]
 
-            # Get groups (pagination handled internally)
-            groups = client.groups.get(filters=filters)
+            # Get journalists (pagination handled internally)
+            journalists = client.journalists.get(filters=filters)
 
             # Iterate over the results without worrying about pagination
-            for g in groups:
-                print(f"ID: {g['group_id']}, Name: {g['name_en']}")
+            for j in journalists:
+                print(f"ID: {j['journalist_id']}, Name: {j['name']}")
             ```
         """
         if filters is None:
@@ -57,7 +57,7 @@ class GroupsResource(BaseResource):
         new_filters.append(limit_filter)
         new_filters.append(skip_filter)
         while True:
-            response = self._client.get_entities(EntityType.GROUP, new_filters)
+            response = self._client.get_entities(EntityType.JOURNALIST, new_filters)
             if response.status != 200:
                 break
             for item in response.json():  # type: ignore
@@ -67,11 +67,25 @@ class GroupsResource(BaseResource):
             new_filters = [f for f in new_filters if f.name != "skip"]
             new_filters.append(skip_filter)
 
-    def get_by_id(self, group_id: int):
-        return self._client.get_entity(EntityType.GROUP, group_id)
+    def get_by_id(self, journalist_id: int):
+        return self._client.get_entity(EntityType.JOURNALIST, journalist_id)
 
-    def update(self, groups: list[dict], filters: list[Filter] | None = None):
-        return self._client.update_entities(EntityType.GROUP, groups, filters)
+    def update(
+        self,
+        journalists: list[dict],
+        filters: list[Filter] | None = None,
+        service: Service = Service.SPARKY,
+    ):
+        return self._client.update_entities(
+            EntityType.JOURNALIST, journalists, filters, service=service
+        )
 
-    def create(self, groups: list[dict], filters: list[Filter] | None = None):
-        return self._client.create_entities(EntityType.GROUP, groups, filters)
+    def create(
+        self,
+        journalists: list[dict],
+        filters: list[Filter] | None = None,
+        service: Service = Service.SPARKY,
+    ):
+        return self._client.create_entities(
+            EntityType.JOURNALIST, journalists, filters, service=service
+        )
