+++ daaily/lucy/resources/project/resource.py
@@ -1,22 +1,21 @@
 from typing import Any, Dict, Generator
 
-from daaily.lucy.enums import EntityType
+from daaily.lucy.enums import EntityType, Service
 from daaily.lucy.models import Filter
 
-from . import BaseResource
+from .. import BaseResource
 
 
-class MaterialsResource(BaseResource):
+class ProjectsResource(BaseResource):
     def get(
         self, filters: list[Filter] | None = None
     ) -> Generator[Dict[str, Any], None, None]:
         """
-        Retrieves materials with optional filtering, returning them as a generator
-        that yields each material one at a time.
+        Retrieves project with optional filtering, returning them as a generator
+        that yields each project one at a time.
 
         Available filters:
-            - manufacturer_id (int): Filter by manufacturer ID.
-            - material_ids (str): Filter by comma separated material IDs.
+            - project_ids (str): Filter by comma separated project IDs.
 
         Note that the following filters are automatically added to the query:
             - skip (int): Number of records to skip.
@@ -26,7 +25,7 @@ class MaterialsResource(BaseResource):
             filters (list[Filter] | None): A list of filters to apply to the query.
 
         Yields:
-            dict: A dictionary representing a single material.
+            dict: A dictionary representing a single project.
 
         Example:
             ```python
@@ -34,11 +33,11 @@ class MaterialsResource(BaseResource):
             filters = [Filter("manufacturer_id", "12345")]
 
             # Get materials (pagination handled internally)
-            materials = client.materials.get(filters=filters)
+            projects = client.projects.get(filters=filters)
 
             # Iterate over the results without worrying about pagination
-            for m in materials:
-                print(f"ID: {m['material_id']}, Name: {m['name_en']}")
+            for p in projects:
+                print(f"ID: {p['project_id']}, Name: {p['name']}")
             ```
         """
         if filters is None:
@@ -58,7 +57,7 @@ class MaterialsResource(BaseResource):
         new_filters.append(limit_filter)
         new_filters.append(skip_filter)
         while True:
-            response = self._client.get_entities(EntityType.MATERIAL, new_filters)
+            response = self._client.get_entities(EntityType.PROJECT, new_filters)
             if response.status != 200:
                 break
             for item in response.json():  # type: ignore
@@ -68,11 +67,25 @@ class MaterialsResource(BaseResource):
             new_filters = [f for f in new_filters if f.name != "skip"]
             new_filters.append(skip_filter)
 
-    def get_by_id(self, material_id: int):
-        return self._client.get_entity(EntityType.MATERIAL, material_id)
+    def get_by_id(self, project_id: int):
+        return self._client.get_entity(EntityType.PROJECT, project_id)
 
-    def update(self, materials: list[dict], filters: list[Filter] | None = None):
-        return self._client.update_entities(EntityType.MATERIAL, materials, filters)
+    def update(
+        self,
+        projects: list[dict],
+        filters: list[Filter] | None = None,
+        service: Service = Service.SPARKY,
+    ):
+        return self._client.update_entities(
+            EntityType.PROJECT, projects, filters, service=service
+        )
 
-    def create(self, materials: list[dict], filters: list[Filter] | None = None):
-        return self._client.create_entities(EntityType.MATERIAL, materials, filters)
+    def create(
+        self,
+        projects: list[dict],
+        filters: list[Filter] | None = None,
+        service: Service = Service.SPARKY,
+    ):
+        return self._client.create_entities(
+            EntityType.PROJECT, projects, filters, service=service
+        )
