+++ daaily/lucy/resources/space/resource.py
@@ -1,21 +1,21 @@
 from typing import Any, Dict, Generator
 
-from daaily.lucy.enums import EntityType
+from daaily.lucy.enums import EntityType, Service
 from daaily.lucy.models import Filter
 
-from . import BaseResource
+from .. import BaseResource
 
 
-class StoriesResource(BaseResource):
+class SpacesResource(BaseResource):
     def get(
         self, filters: list[Filter] | None = None
     ) -> Generator[Dict[str, Any], None, None]:
         """
-        Retrieves stories with optional filtering, returning them as a generator
-        that yields each story one at a time.
+        Retrieves spaces with optional filtering, returning them as a generator
+        that yields each space one at a time.
 
         Available filters:
-            - story_ids (str): Filter by comma separated story IDs.
+            - space_ids (str): Filter by comma separated space IDs.
 
         Note that the following filters are automatically added to the query:
             - skip (int): Number of records to skip.
@@ -25,19 +25,19 @@ class StoriesResource(BaseResource):
             filters (list[Filter] | None): A list of filters to apply to the query.
 
         Yields:
-            dict: A dictionary representing a single story.
+            dict: A dictionary representing a single space.
 
         Example:
             ```python
             # Define filters
-            filters = [Filter("story_ids", "12345, 78910")]
+            filters = [Filter("space_ids", "12345, 78910")]
 
-            # Get stories (pagination handled internally)
-            stories = client.stories.get(filters=filters)
+            # Get spaces (pagination handled internally)
+            spaces = client.spaces.get(filters=filters)
 
             # Iterate over the results without worrying about pagination
-            for s in stories:
-                print(f"ID: {s['story_id']}, Name: {s['name']}")
+            for s in spaces:
+                print(f"ID: {s['space_id']}, Space Type: {s['space_type']}")
             ```
         """
         if filters is None:
@@ -57,7 +57,7 @@ class StoriesResource(BaseResource):
         new_filters.append(limit_filter)
         new_filters.append(skip_filter)
         while True:
-            response = self._client.get_entities(EntityType.STORY, new_filters)
+            response = self._client.get_entities(EntityType.SPACE, new_filters)
             if response.status != 200:
                 break
             for item in response.json():  # type: ignore
@@ -67,11 +67,25 @@ class StoriesResource(BaseResource):
             new_filters = [f for f in new_filters if f.name != "skip"]
             new_filters.append(skip_filter)
 
-    def get_by_id(self, story_id: int):
-        return self._client.get_entity(EntityType.STORY, story_id)
+    def get_by_id(self, space_id: int):
+        return self._client.get_entity(EntityType.SPACE, space_id)
 
-    def update(self, stories: list[dict], filters: list[Filter] | None = None):
-        return self._client.update_entities(EntityType.STORY, stories, filters)
+    def update(
+        self,
+        spaces: list[dict],
+        filters: list[Filter] | None = None,
+        service: Service = Service.SPARKY,
+    ):
+        return self._client.update_entities(
+            EntityType.SPACE, spaces, filters, service=service
+        )
 
-    def create(self, stories: list[dict], filters: list[Filter] | None = None):
-        return self._client.create_entities(EntityType.STORY, stories, filters)
+    def create(
+        self,
+        spaces: list[dict],
+        filters: list[Filter] | None = None,
+        service: Service = Service.SPARKY,
+    ):
+        return self._client.create_entities(
+            EntityType.SPACE, spaces, filters, service=service
+        )
