+++ daaily/lucy/resources/story/resource.py
@@ -1,21 +1,21 @@
 from typing import Any, Dict, Generator
 
-from daaily.lucy.enums import EntityType
+from daaily.lucy.enums import EntityType, Service
 from daaily.lucy.models import Filter
 
-from . import BaseResource
+from .. import BaseResource
 
 
-class JournalistsResource(BaseResource):
+class StoriesResource(BaseResource):
     def get(
         self, filters: list[Filter] | None = None
     ) -> Generator[Dict[str, Any], None, None]:
         """
-        Retrieves journalists with optional filtering, returning them as a generator
-        that yields each journalist one at a time.
+        Retrieves stories with optional filtering, returning them as a generator
+        that yields each story one at a time.
 
         Available filters:
-            - journalist_ids (str): Filter by comma separated journalist IDs.
+            - story_ids (str): Filter by comma separated story IDs.
 
         Note that the following filters are automatically added to the query:
             - skip (int): Number of records to skip.
@@ -25,19 +25,19 @@ class JournalistsResource(BaseResource):
             filters (list[Filter] | None): A list of filters to apply to the query.
 
         Yields:
-            dict: A dictionary representing a single journalist.
+            dict: A dictionary representing a single story.
 
         Example:
             ```python
             # Define filters
-            filters = [Filter("journalist_ids", "12345, 12322")]
+            filters = [Filter("story_ids", "12345, 78910")]
 
-            # Get journalists (pagination handled internally)
-            journalists = client.journalists.get(filters=filters)
+            # Get stories (pagination handled internally)
+            stories = client.stories.get(filters=filters)
 
             # Iterate over the results without worrying about pagination
-            for j in journalists:
-                print(f"ID: {j['journalist_id']}, Name: {j['name']}")
+            for s in stories:
+                print(f"ID: {s['story_id']}, Name: {s['name']}")
             ```
         """
         if filters is None:
@@ -57,7 +57,7 @@ class JournalistsResource(BaseResource):
         new_filters.append(limit_filter)
         new_filters.append(skip_filter)
         while True:
-            response = self._client.get_entities(EntityType.JOURNALIST, new_filters)
+            response = self._client.get_entities(EntityType.STORY, new_filters)
             if response.status != 200:
                 break
             for item in response.json():  # type: ignore
@@ -67,11 +67,25 @@ class JournalistsResource(BaseResource):
             new_filters = [f for f in new_filters if f.name != "skip"]
             new_filters.append(skip_filter)
 
-    def get_by_id(self, journalist_id: int):
-        return self._client.get_entity(EntityType.JOURNALIST, journalist_id)
+    def get_by_id(self, story_id: int):
+        return self._client.get_entity(EntityType.STORY, story_id)
 
-    def update(self, journalists: list[dict], filters: list[Filter] | None = None):
-        return self._client.update_entities(EntityType.JOURNALIST, journalists, filters)
+    def update(
+        self,
+        stories: list[dict],
+        filters: list[Filter] | None = None,
+        service: Service = Service.SPARKY,
+    ):
+        return self._client.update_entities(
+            EntityType.STORY, stories, filters, service=service
+        )
 
-    def create(self, journalists: list[dict], filters: list[Filter] | None = None):
-        return self._client.create_entities(EntityType.JOURNALIST, journalists, filters)
+    def create(
+        self,
+        stories: list[dict],
+        filters: list[Filter] | None = None,
+        service: Service = Service.SPARKY,
+    ):
+        return self._client.create_entities(
+            EntityType.STORY, stories, filters, service=service
+        )
